Single Motor Lift works (but the motor is too wimpy)
