+++ Sixteen750/src/main/java/org/firstinspires/ftc/sixteen750/command/autonomous/blue_away/AutoBlueAwayConeStackCommand.java
@@ -4,7 +4,6 @@ import org.firstinspires.ftc.sixteen750.command.claw.ClawCloseCommand;
 import org.firstinspires.ftc.sixteen750.command.claw.ClawOpenCommand;
 import org.firstinspires.ftc.sixteen750.subsystem.ClawSubsystem;
 import org.firstinspires.ftc.sixteen750.subsystem.LiftSubsystem;
-import org.firstinspires.ftc.sixteen750.subsystem.LiftSubsystem2;
 
 import com.technototes.library.command.SequentialCommandGroup;
 import com.technototes.path.subsystem.MecanumDrivebaseSubsystem;
