+++ Sixteen750/src/main/java/org/firstinspires/ftc/sixteen750/command/lift/LiftCarryCommand.java
@@ -13,6 +13,6 @@ public class LiftCarryCommand implements Command {
 
     @Override
     public void execute() {
-        liftSubsystem.gotoGroundJunction();
+        liftSubsystem.groundJunction();
     }
 }
