+++ Sixteen750/src/main/java/org/firstinspires/ftc/sixteen750/command/lift/LiftGroundJunctionCommand.java
@@ -13,6 +13,6 @@ public class LiftGroundJunctionCommand implements Command {
 
     @Override
     public void execute() {
-        liftSubsystem.gotoGroundJunction();
+        liftSubsystem.groundJunction();
     }
 }
