+++ Sixteen750/src/main/java/org/firstinspires/ftc/sixteen750/command/lift/LiftHighPoleCommand.java
@@ -13,6 +13,6 @@ public class LiftHighPoleCommand implements Command {
 
     @Override
     public void execute() {
-        liftSubsystem.gotoHighPole();
+        liftSubsystem.highPole();
     }
 }
