+++ Sixteen750/src/main/java/org/firstinspires/ftc/sixteen750/command/lift/LiftIntakeCommand.java
@@ -13,6 +13,6 @@ public class LiftIntakeCommand implements Command {
 
     @Override
     public void execute() {
-        liftSubsystem.gotoGroundJunction();
+        liftSubsystem.intakePos();
     }
 }
