+++ Sixteen750/src/main/java/org/firstinspires/ftc/sixteen750/command/lift/LiftLowPoleCommand.java
@@ -13,6 +13,6 @@ public class LiftLowPoleCommand implements Command {
 
     @Override
     public void execute() {
-        liftSubsystem.gotoLowPole();
+        liftSubsystem.lowPole();
     }
 }
