+++ Sixteen750/src/main/java/org/firstinspires/ftc/sixteen750/command/lift/LiftMidPoleCommand.java
@@ -13,6 +13,6 @@ public class LiftMidPoleCommand implements Command {
 
     @Override
     public void execute() {
-        liftSubsystem.gotoMidPole();
+        liftSubsystem.midPole();
     }
 }
