--- Sixteen750/src/main/java/org/firstinspires/ftc/sixteen750/command/lift2/Lift2LowJunction.java
@@ -1,17 +1,0 @@
-package org.firstinspires.ftc.sixteen750.command.lift2;
-
-import com.technototes.library.command.Command;
-
-import org.firstinspires.ftc.sixteen750.subsystem.LiftSubsystem2;
-
-public class Lift2LowJunction implements Command {
-    private LiftSubsystem2 liftSubsystem2;
-
-    public Lift2LowJunction(LiftSubsystem2 l2){
-        liftSubsystem2 = l2;
-    }
-    @Override
-    public void execute() {
-        liftSubsystem2.lowPole();
-    }
-}
